unify test harness: type-check + execute by default
FormalityTest.ok() now type-checks AND executes (codegen + MiniRust)
by default. Tests without a main function must call .skip_execute().
Tests can assert stdout with .expect_output("...").

- Move codegen tests from crates/formality-rust/src/codegen/test.rs
  to tests/codegen.rs using the unified harness
- Fix reference tests to use exists<'a> instead of 'static
  (required now that type-checking runs before codegen)
- Add minirust-rs and libspecr as root crate dependencies
- Add .skip_execute() to all existing integration tests (no main)
