+++ src/test_util.rs
@@ -19,6 +19,8 @@ enum BackendExpect {
 pub struct FormalityTest {
     input: String,
     rustc_override: Option<BackendExpect>,
+    skip_execute: bool,
+    expected_output: Option<String>,
 }
 
 impl FormalityTest {
@@ -26,7 +28,22 @@ impl FormalityTest {
         Self {
             input: input.into(),
             rustc_override: None,
+            skip_execute: false,
+            expected_output: None,
         }
+    }
+
+    /// Skip execution (codegen + run) even if the program type-checks.
+    /// Use for programs that have no `main` or test features codegen doesn't support yet.
+    pub fn skip_execute(mut self) -> Self {
+        self.skip_execute = true;
+        self
+    }
+
+    /// Assert that execution produces the given stdout output.
+    pub fn expect_output(mut self, output: impl Into<String>) -> Self {
+        self.expected_output = Some(output.into());
+        self
     }
 
     /// Require rustc to accept the program (always runs rustc).
@@ -41,17 +58,28 @@ impl FormalityTest {
         self
     }
 
-    /// Assert formality accepts the program. Also runs rustc if overridden,
-    /// or if `FORMALITY_RUN_RUSTC=1` (in which case rustc must also accept).
+    /// Assert formality accepts the program. After type-checking passes,
+    /// also runs codegen + execution unless `.skip_execute()` was called.
+    /// Also runs rustc if overridden or if `FORMALITY_RUN_RUSTC=1`.
     #[track_caller]
     pub fn ok(self) {
         let Self {
             input,
             rustc_override,
+            skip_execute,
+            expected_output,
         } = self;
 
         let proof_tree = test_program_ok(&input).expect("expected program to pass");
         formality_core::judgment::coverage::record_coverage(std::iter::once(&proof_tree));
+
+        if !skip_execute {
+            let stdout = execute_program(&input);
+            if let Some(expected) = expected_output {
+                assert_eq!(stdout, expected, "program output mismatch");
+            }
+        }
+
         if let Some(expect) = rustc_override {
             run_rustc_backend(&input, expect);
         } else if run_rustc() {
@@ -66,6 +94,8 @@ impl FormalityTest {
         let Self {
             input,
             rustc_override,
+            skip_execute: _,
+            expected_output: _,
         } = self;
 
         test_program_ok(&input).assert_err_leaves(expect);
@@ -74,6 +104,66 @@ impl FormalityTest {
             run_rustc_backend(&input, rustc);
         }
     }
+}
+
+#[track_caller]
+fn execute_program(input: &str) -> String {
+    let crates: formality_rust::grammar::Crates =
+        formality_rust::rust::try_term(input).expect("failed to parse program");
+
+    let has_main = crates.crates.iter().any(|c| {
+        c.items.iter().any(|item| {
+            matches!(item, formality_rust::grammar::CrateItem::Fn(f) if &**f.id == "main")
+        })
+    });
+
+    if !has_main {
+        panic!(
+            "program has no `main` function — add one, or call `.skip_execute()` on the test"
+        );
+    }
+
+    let program =
+        formality_rust::codegen::codegen_program(&crates).expect("codegen failed");
+
+    let stdout_buf = std::sync::Arc::new(std::sync::Mutex::new(Vec::<u8>::new()));
+    let stderr_buf = std::sync::Arc::new(std::sync::Mutex::new(Vec::<u8>::new()));
+
+    let stdout = libspecr::DynWrite::new(SharedWriter(stdout_buf.clone()));
+    let stderr = libspecr::DynWrite::new(SharedWriter(stderr_buf.clone()));
+
+    type Mem = minirust_rs::mem::BasicMemory<minirust_rs::prelude::x86_64>;
+    let mut machine: minirust_rs::lang::Machine<Mem> =
+        minirust_rs::lang::Machine::new(program, stdout, stderr)
+            .get_internal()
+            .expect("machine creation failed");
+
+    loop {
+        match machine.step().get_internal() {
+            Ok(()) => continue,
+            Err(minirust_rs::prelude::TerminationInfo::MachineStop) => break,
+            Err(e) => panic!("execution error: {e:?}"),
+        }
+    }
+
+    let bytes = stdout_buf.lock().unwrap().clone();
+    String::from_utf8(bytes).expect("stdout was not valid UTF-8")
+}
+
+struct SharedWriter(std::sync::Arc<std::sync::Mutex<Vec<u8>>>);
+
+impl std::io::Write for SharedWriter {
+    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
+        self.0.lock().unwrap().extend_from_slice(buf);
+        Ok(buf.len())
+    }
+    fn flush(&mut self) -> std::io::Result<()> {
+        Ok(())
+    }
+}
+
+impl libspecr::hidden::GcCompat for SharedWriter {
+    fn points_to(&self, _m: &mut std::collections::HashSet<usize>) {}
 }
 
 #[track_caller]
