+++ tests/coherence_orphan.rs
@@ -80,7 +80,7 @@ fn mirror_FooStruct() {
         struct FooStruct {}
         impl CoreTrait for <FooStruct as Mirror>::Assoc {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -93,7 +93,7 @@ fn covered_VecT() {
         struct FooStruct {}
         impl<T> CoreTrait<FooStruct> for Vec<T> {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -190,5 +190,5 @@ fn CoreTraitLocal_for_AliasToKnown_in_Foo() {
         struct FooStruct {}
         impl CoreTrait<FooStruct> for <() as Unit>::Assoc {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
