+++ tests/coherence_overlap.rs
@@ -143,7 +143,7 @@ fn neg_CoreTrait_for_CoreStruct_implies_no_overlap() {
         impl<T> FooTrait for T where T: CoreTrait {}
         impl FooTrait for CoreStruct {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -201,7 +201,7 @@ fn u32_T_where_T_Not_impls() {
 
         trait Not {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -223,7 +223,7 @@ fn u32_i32_impls() {
         impl Foo for u32 {}
         impl Foo for i32 {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -300,7 +300,7 @@ fn is_local_unknowable_trait_ref() {
             <T as Project>::Assoc: Foo<U> {}
         impl<T> Overlap<LocalType> for () {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
