+++ tests/const_generics_rv_tsv_parse.rs
@@ -7,5 +7,5 @@ fn parse_minirust_22() {
         trait Trait<const N> where type_of_const N is usize {}
         impl Trait<usize(22)> for u32 {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
