+++ tests/consts.rs
@@ -7,7 +7,7 @@ fn nonsense_rigid_const_bound() {
         // This where-clause is not *provable*, but it is well-formed.
         trait Foo where type_of_const true is u32 {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -18,7 +18,7 @@ fn ok() {
 
         impl<const C> Foo<C> for u32 where type_of_const C is bool {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -43,7 +43,7 @@ fn holds() {
 
         impl Foo<true> for u32 {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -51,7 +51,7 @@ fn rigid_const_bound() {
     FormalityTest::new(crates![crate Foo {
         trait Foo where type_of_const true is bool {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -84,7 +84,7 @@ fn generic_match() {
         // Here, the impl matches the trait bound.
         impl<const C> Foo<C> for u32 where type_of_const C is bool {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -92,5 +92,5 @@ fn multiple_type_of_const() {
     FormalityTest::new(crates![crate Foo {
         trait Foo<const C> where type_of_const C is bool, type_of_const C is u32 {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
