+++ tests/decl_safety.rs
@@ -8,7 +8,7 @@ fn unsafe_trait() {
         unsafe trait Foo {}
         unsafe impl Foo for u32 {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -17,7 +17,7 @@ fn safe_trait() {
         safe trait Foo {}
         safe impl Foo for u32 {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
@@ -26,7 +26,7 @@ fn unsafe_trait_negative_impl() {
         unsafe trait Foo {}
         impl !Foo for u32 {}
     }])
-    .ok()
+    .skip_execute().ok()
 }
 
 #[test]
